250811 trying assignements on my own.
encountered issues: assignment_4.html : there seems to be a margin collapsing occured or something unknown that cause the white background top-border-line is in touch with the black background (body tag). The white box (container class) needs to be gapped a little below the background (body tag) in order to have some space in-between. However no where to find to fix this issue.